+++ Code/Common/itkOneWayEquivalencyTable.cxx
@@ -51,7 +51,7 @@ void OneWayEquivalencyTable::Flatten()
     }
 }
 
-unsigned long OneWayEquivalencyTable::RecursiveLookup(const unsigned a) const
+unsigned long OneWayEquivalencyTable::RecursiveLookup(const unsigned long a) const
 {
   unsigned long ans = a;
   unsigned long last_ans=a;
